move LocalSecuritySettingContents to under Admin/Security
