+++ src/components/Admin/Security/LocalSecuritySettingContents.jsx
@@ -3,12 +3,12 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { withTranslation } from 'react-i18next';
 
-import { withUnstatedContainers } from '../../UnstatedUtils';
-import { toastSuccess, toastError } from '../../../util/apiNotification';
+import { withUnstatedContainers } from '../../../client/js/components/UnstatedUtils';
+import { toastSuccess, toastError } from '../../../client/js/util/apiNotification';
 
-import AppContainer from '../../../services/AppContainer';
-import AdminGeneralSecurityContainer from '../../../services/AdminGeneralSecurityContainer';
-import AdminLocalSecurityContainer from '../../../services/AdminLocalSecurityContainer';
+import AppContainer from '../../../client/js/services/AppContainer';
+import AdminGeneralSecurityContainer from '../../../client/js/services/AdminGeneralSecurityContainer';
+import AdminLocalSecurityContainer from '../../../client/js/services/AdminLocalSecurityContainer';
 
 class LocalSecuritySettingContents extends React.Component {
